Additional movement
This commit moves some functions and interface implementations into the
utils directory.  This should complete the separation between the bare
parser framework and its common code, and utilities intended to aid in the
construction of parsers using the ptk package.  This commit also corrects
some minor errors in the documentation, as well as documenting the new
locations of some of the referenced elements.