Add Parser, application state to State
This commit adds the parser and an application state to the State type.
It also includes changes to the Literal, Prefix, Infix, and InfixR helper
functions to pass the state along to the factories.  These changes reduce
the number of arguments passed to the parse functions by allowing the
parser to be pulled from the state, as well as giving the application
using the toolkit the opportunity to include its own application state, if
that is needed for the parser being built.