Add Location.Incr
This commit contains the common/character.go file, describing a character
that will be used by the lexer.  This file also contains the EOF
character, a special character that is then referenced by the
FileLocation.Incr routine, which implements the new Incr function added
to the Location interface; Incr returns the location of the character
passed to it.  This will be used by the lexer's scanner to produce
locations as it returns characters to the lexer.