Create the scanner
This commit contains the scanner.  A scanner is a CharStream that returns
consecutive runes from an io.Reader object.  The scanner is also
responsible for handling line endings; this is done through the LineStyle
type, of which four are available: UNIXLineStyle, MacLineStyle,
DOSLineStyle, and UnknownLineStyle, which guesses the line ending style
from the contents.  The scanner can use a kent.Reporter object for
reporting an encoding error in the input, but it simply returns a
utf8.RuneError rune for such encoding errors and continues processing.
Any read errors are reported at the end of the scan when returning the
common.EOF to indicate end of file.