Create BackTracker
This commit creates BackTracker, a CharStream implementation that wraps
another CharStream and allows for easy backtracking using a list.List.
Each character is saved to a list, and calling BackTrack will cause the
Next method to restart at the beginning of the list of saved characters.
A maximum number of characters to save may be set; also, up to a certain
number of characters may be accepted and discarded from the set of saved
characters.