Specify the lexer interface
This commit brings in the four main interfaces that define a lexer, along
with mocks and unit tests for the mocks.  These interfaces are the Lexer,
which is the core object; the State, which maintains state during the
lexing; the Classifier, which represents application code that selects a
Recognizer, which is additional application code that actually extracts
characters from a character stream and generates tokens from them.