Create the lexical analyzer
This commit contains the lexical analyzer framework.  A user instantiates
a Lexer object (using New) and passes it a default classifier, then calls
the Lex method with a CharStream.  This then returns a common.TokenStream
(a lexer.State under the covers) that yields tokens until the character
stream is exhausted.