Create an argument character stream
This commit includes NewArgumentCharStream, which brings together the
memoizing character stream, chaining character stream, and the scanner to
create a character stream that iterates over the characters in a list of
arguments, with a joiner string logically placed between each argument.
This will enable easily parsing an expression provided on a command line
without having to either use strings.Join or require the user to surround
the expression with double quotes.