Properly detect EOF from source
While developing a utility recognizer, it was realized that the lexer
state did not have a reliable way of determining if there were more
characters to retrieve from the character stream.  Such a call cannot be
added to the ordinary CharStream, because not all versions know that
without trying to return a character, so this commit adds a More method to
the BackTracker type and updates lexer.State to utilize this method.  This
will ensure that lexer.State.Next does not attempt to call a recognizer
unless there is input for it to consume, and the recognizer need do
nothing special to handle EOF.