Initial tweaks to Makefile
This commit contains some initial tweaks to the Makefile.  One of these
tweaks is to simply remove the copyright notice, since it's not needed for
this file.  A second is to create the cover-report target to generate the
coverage report.  The only substantive change is to add the
"-coverpkg=./..." argument to the call to "go test", which ensures that
coverage results include calls in other packages; this is necessary for,
say, functional tests divorced from the rest of the code to generate
correct coverage data.