Scanner refactoring
This commit contains the beginning of the refactoring of the ptk package
by adding a scanner subpackage and "moving" (for now, copying) of various
components that have been placed elsewhere in the ptk repository.  The
concept of a "scanner" has replaced the concept of a CharStream, and the
lexer will soon be updated accordingly.

This commit also represents a change in methodology, where the underlying
types are exported, functions return concrete types, and mocks are no
longer exported.  All of this represents a maturing of my development
philosophy in go.