Add an imports target
One of the nice things about goimports which golangci-lint cannot do is
that goimports is able to manage the imports for a package, adding newly
needed imports to a source file.  It's even able to make a reasonable
guess as to which import may be needed for a referenced package and add
that package import to go.mod.  Since this cannot be done by
golangci-lint, this commit adds back an "imports" target to the Makefile
that simply runs goimports on the source files.  This can be used during
development, prior to running a "make test", to manage the imports.