Move Token into lexer
This commit kicks off the lexer refactor series by moving the Token and
TokenStream definitions into the lexer.  This entailed a lot of other
changes, including moving the soon-to-be-excised Stack code from common
into internal and creating a TokenNode wrapper since a Token cannot
implement the self-referential Node interface anymore.