Refactor parser
This commit refactors the parser by converting the State to an object that
is passed into the New function, which returns a Parser that is an
implementation of IParser.  This also eliminates the need for the Stack
object, so that code has been eliminated.  A PushBackLexer type has also
been created that serves a similar purpose as the BackTracker type from
the lexer.