Allow for custom BackTracker
The parser.New function implements selective wrapping of the lexer, only
wrapping it in a PushBackLexer if it doesn't implement the IPushBackLexer
interface.  This allows callers to substitute their own IPushBackLexer
implementation.  This commit extends this logic into lexer.New, allowing
alternate implementations of IBackTracker to be passed when constructing
the Lexer.  This will allow users to extend the functionality if desired.